Extract match scoring logic into MatchScoring service
