+++ lambdas/src/main/kotlin/scorcerer/server/schedule/MatchStarter.kt
@@ -8,8 +8,8 @@ import org.jetbrains.exposed.v1.jdbc.transactions.transaction
 import org.openapitools.server.models.Match
 import scorcerer.server.db.tables.MatchTable
 import scorcerer.server.log
-import scorcerer.server.resources.getMatchDay
-import scorcerer.server.resources.setScore
+import scorcerer.server.services.getMatchDay
+import scorcerer.server.services.setScore
 import scorcerer.utils.LeaderboardS3Service
 import java.time.Clock
 import java.time.OffsetDateTime
